Refactor FuncPtr
