+++ src/interface.rs
@@ -1,2 +1,3 @@
+mod func_ptr;
 pub mod injector;
 mod verifier;
