+++ src/interface/injector.rs
@@ -1,10 +1,10 @@
 use crate::injector_core::common::*;
 use crate::injector_core::internal::*;
+pub use crate::interface::func_ptr::FuncPtr;
 pub use crate::interface::verifier::CallCountVerifier;
 
 use std::future::Future;
 use std::pin::Pin;
-use std::ptr::NonNull;
 use std::sync::Mutex;
 use std::sync::MutexGuard;
 use std::task::Context;
@@ -494,44 +494,6 @@ macro_rules! fake {
          let raw_ptr = f as *const ();
          (unsafe { FuncPtr::new(raw_ptr, std::any::type_name_of_val(&f)) }, verifier)
     }};
-}
-
-/// A safe wrapper around a raw function pointer.
-///
-/// `FuncPtr` encapsulates a non-null function pointer and provides safe
-/// creation and access methods. It's used throughout injectorpp
-/// to represent both original functions to be mocked and their replacement
-/// implementations.
-///
-/// # Safety
-///
-/// The caller must ensure that the pointer is valid and points to a function.
-pub struct FuncPtr {
-    /// The internal representation of the function pointer.
-    ///
-    /// This is a wrapper around a non-null pointer to ensure safety.
-    func_ptr_internal: FuncPtrInternal,
-    signature: &'static str,
-}
-
-impl FuncPtr {
-    /// Creates a new `FuncPtr` from a raw pointer.
-    ///
-    /// # Safety
-    ///
-    /// The caller must ensure that the pointer is valid and points to a function.
-    pub unsafe fn new(ptr: *const (), signature: &'static str) -> Self {
-        // While these basic checks are performed, it is not a substitute for
-        // proper function pointer validation. The caller must ensure that the
-        // pointer is indeed a valid function pointer.
-        let p = ptr as *mut ();
-        let nn = NonNull::new(p).expect("Pointer must not be null");
-
-        Self {
-            func_ptr_internal: FuncPtrInternal::new(nn),
-            signature,
-        }
-    }
 }
 
 /// A high-level type that holds patch guards so that when it goes out of scope,
